GameOver Screen Score text & Optimization on music files
