+++ Logic/Logic.vcxproj
@@ -205,16 +205,17 @@
     <ClInclude Include="include\AI\EnemySoarer.h" />
     <ClInclude Include="include\AI\EnemyTotem.h" />
     <ClInclude Include="include\Misc\Fader.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuCards.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuControls.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuFirstTime.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuCards.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuControls.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuFirstTime.h" />
     <ClInclude Include="include\Misc\GUI\iMenuFX.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuHighscore.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuIntro.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuLoading.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuSettings.h" />
-    <ClInclude Include="include\Misc\GUI\iMenuSkillPick.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuHighscore.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuIntro.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuLoading.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuSettings.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuSkillPick.h" />
     <ClInclude Include="include\Misc\GUI\Slider.h" />
+    <ClInclude Include="include\Misc\GUI\Specific\iMenuGameOver.h" />
     <ClInclude Include="include\Misc\Network\dbConnect.h" />
     <ClInclude Include="include\Misc\Test\TestingWork.h" />
     <ClInclude Include="include\Physics\FuncContactResult.h" />
@@ -322,17 +323,18 @@
     <ClCompile Include="source\AI\EnemySoarer.cpp" />
     <ClCompile Include="source\AI\EnemyTotem.cpp" />
     <ClCompile Include="source\Misc\GUI\iMenuAction.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuCards.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuControls.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuCards.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuControls.cpp" />
     <ClCompile Include="source\Misc\GUI\iMenuFactory.cpp" />
     <ClCompile Include="source\Misc\GUI\iMenu.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuFirstTime.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuHighscore.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuIntro.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuLoading.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuFirstTime.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuGameOver.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuHighscore.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuIntro.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuLoading.cpp" />
     <ClCompile Include="source\Misc\GUI\iMenuMachine.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuSettings.cpp" />
-    <ClCompile Include="source\Misc\GUI\iMenuSkillPick.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuSettings.cpp" />
+    <ClCompile Include="source\Misc\GUI\Specific\iMenuSkillPick.cpp" />
     <ClCompile Include="source\Misc\Network\dbConnect.cpp" />
     <ClCompile Include="source\Misc\GUI\Slider.cpp" />
     <ClCompile Include="source\Misc\Test\TestingWork.cpp" />
